pyproject.toml
[tool.poetry]
name = "fastapi-tdd-docker"
version = "0.1.0"
description = ""
authors = ["Mihlos <miguel.mish@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
asyncpg = "^0.24.0"
tortoise-orm = "^0.17.8"
aerich = "^0.5.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
